Move run_metrics
